command console tools

+++ gdb/mod_command_stack.c
@@ -6,12 +6,12 @@
 #define COMMAND_STACK_LOGSIZE 3
 #endif
 #define COMMAND_STACK_MASK ((1<<COMMAND_STACK_LOGSIZE)-1)
-uint32_t command_stack[COMMAND_STACK_MASK+1];
-uint32_t command_stack_top;
-void command_stack_push(uint32_t word) {
+uintptr_t command_stack[COMMAND_STACK_MASK+1];
+uintptr_t command_stack_top;
+void command_stack_push(uintptr_t word) {
     command_stack[(--command_stack_top) & COMMAND_STACK_MASK] = word;
 }
-uint32_t command_stack_pop(void) {
+uintptr_t command_stack_pop(void) {
     return command_stack[(command_stack_top++) & COMMAND_STACK_MASK];
 }
